backend/requirements.txt
bandit>=1.7.0 
certifi==2022.6.15
chardet==3.0.4 
coreapi==2.3.3 
coreschema==0.0.4 
django==5.2.6
django-debug-toolbar>=4.4.5
# psycopg2>=2.8.4
psycopg>=3.1.8
django-cors-headers==4.1.0 
django-rest-swagger==2.2.0
drf-spectacular==0.27.2
django-templated-mail==1.1.1 
djangorestframework>=3.9.3 
djangorestframework-simplejwt>=4.4.0 
djoser==2.3.3
idna==2.8 
itypes==1.1.0 
# Jinja2==2.10.1
MarkupSafe==1.1.1 
openapi-codec==1.3.2 
PyJWT==2.5.0
pytz==2019.1 
requests==2.32.4
simplejson==3.16.0 
sqlparse>=0.3.1 
uritemplate==3.0.0 
# psycopg2-binary==2.8.4 

# latest compatible version of urllib3 
urllib3==2.2.0

# Dev dependencies 
ptvsd==4.3.2